requirements.txt
dash==3.3.0
dash-bootstrap-components==2.0.4
dash-core-components~=2.0
coverage~=7.3
dash-iconify~=0.1
dash-mantine-components~=2.4.0
langchain==1.2.0
langgraph==1.0.5
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
markdownify~=0.11
mypy~=1.5
numpy~=1.26
pandas~=2.1
pytest~=7.4
pytest-cov~=4.1
python-dotenv~=1.0
sentence-transformers==5.2.0
tiktoken~=0.5
gunicorn~=21.2
waitress~=2.1
pysqlite3-binary~=0.5
langgraph-checkpoint-postgres==3.0.3
#"psycopg[binary]" #Need to confirm this will install via requirements.txt